Small changes, moving client manager
